% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fintb-package.R
\docType{package}
\name{findtb-package}
\alias{findtb}
\alias{findtb-package}
\title{findtb: Download and Clean WHO TB Data}
\description{
To learn more about findtb, see the individual help files for each dataset.
}
\author{
\strong{Maintainer}: Anna Mantsoki \email{anna.mantsoki@finddx.org}

Authors:
\itemize{
  \item Mike Page \email{mike@cynkra.com}
  \item Andrea Ranzato \email{andrea@cynkra.com}
}

Other contributors:
\itemize{
  \item FIND [copyright holder, funder]
}

}
\keyword{internal}
