% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{tb}
\alias{tb}
\alias{tb_budget}
\alias{tb_community}
\alias{tb_estimates}
\alias{tb_expenditures}
\alias{tb_labs}
\alias{tb_notifications}
\alias{tb_all}
\title{World Health Organization TB data}
\format{
A data frame with 6 columns and a variable number of rows:
\describe{
\item{country}{Country name}
\item{country_code}{3 letter ISO country codes}
\item{g_whoregion}{WHO region}
\item{year}{Year}
\item{variable}{see \code{tb_dictionary} for definitions}
\item{value}{value}
}
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
\description{
This package exports seven TB data sets from the World Health Organization
Global Tuberculosis Report:
\itemize{
\item budget
\item community
\item estimates
\item expenditures
\item labs
\item notifications
\item all (all of the above combined)
}
}
