% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{tb_all}
\alias{tb_all}
\alias{tb_budget}
\alias{tb_community}
\alias{tb_estimates}
\alias{tb_expenditure}
\alias{tb_labs}
\alias{tb_notifications}
\title{World Health Organization TB data}
\format{
A data frame with 6 columns and a variable number of rows:
\describe{
\item{country}{Country name}
\item{country_code}{3 letter ISO country codes}
\item{g_whoregion}{WHO region}
\item{year}{Year}
\item{variable}{see \code{tb_dictionary} for definitions}
\item{value}{value}
}
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
\description{
This package exports seven TB data sets from the World Health Organization
Global Tuberculosis Report:
\itemize{
\item tb_all (all of the below combined)
\item tb_budget
\item tb_community
\item tb_estimates
\item tb_expenditures
\item tb_labs
\item tb_notifications
}
}
