% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tb_dictionary}
\alias{tb_dictionary}
\title{World Health Organization TB dictionary}
\format{
A data frame with 45 rows and 4 columns:
\describe{
\item{dataset}{Matching tb_* dataset}
\item{variable}{Matching variable from tb_* dataset}
\item{definition}{Definition}
}
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
\usage{
tb_dictionary
}
\description{
A dictionary of defintions to match to the tb_* family of exported data sets.
}
\keyword{datasets}
