% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tb_high_burden}
\alias{tb_high_burden}
\title{World Health Organization high TB burden countries}
\format{
A data frame with 180 rows and 3 columns:
\describe{
\item{country}{Country name}
\item{year}{Year}
\item{share_global_incidence}{Share of the global TB burden (in terms of the absolute number of incident cases)}
}
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
\usage{
tb_high_burden
}
\description{
A list of high burden TB countries.
}
\keyword{datasets}
