% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{who}
\alias{who}
\title{Data from WHO}
\description{
The data sets currently available from WHO in this package are:
\itemize{
\item budget
\item community
\item estimates
\item expenditures
\item labs
\item notifications
}
}
